feat(ppm): Allow switching to a new interval size for new partitions
The provisioning applies the following method:

- Find the currently existing range [start,end]. No gap is allowed.
Only the bounds of the partitions are considered. The names are
ignored.

- Process "retention" by creating missing partitions before
"start", sized with the currently configured "interval".

- Process "preProvisioned" by creating missing partitions after
"end", sized with the currently configured "interval".

For both "retention" and "preProvisioned", add a shorter partition
when necessary, if the range to create does not divide evenly by
"interval". Short partitions are suffixed with YYYYMMDD_YYYYMMDD
representing the partition bounds.

Signed-off-by: Daniel Vérité <dverite@gmail.com>