fix: do not rotate timestamptz to UTC
In order to be timezone-agnostic, we submit partition bounds without
timezone offsets, like:
 VALUES FROM ('YYYY-MM-DD 00:00:00') TO ('YYYY-MM-DD 00:00:00')
The time zone of the session is automatically applied by Postgres.

When reading back these values, we get for instance
 YYYY-MM-DD 00:00:00+02
if the current time zone offset is +02.

If we rotate this timestamp locally to UTC and resubmit it without the
offset, it is now incorrect (unless the server happens to be
configured in UTC).

The fix consists of ignoring the offset when reading the current
bounds, to be consistently timezone-agnostic. The offset always
corresponds implicitly to the time zone of the SQL session, and
must not be forced in the value of the timestamp.

The SQL time zone can be changed dynamically with the PGTZ environment
variable if desired.

Signed-off-by: Daniel Vérité <dverite@gmail.com>